refactor(eo, connections): update structure / architecture to follow our guidelines
