+++ libs/eo/connections/data-access-api/index.ts
@@ -1,0 +1,2 @@
+export * from './connections.service';
+export * from './invite-connection.service';
