+++ libs/eo/connections/data-access-api/project.json
@@ -1,0 +1,8 @@
+{
+  "name": "eo-connections-data-access-api",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "sourceRoot": "libs/eo/connections/data-access-api",
+  "prefix": "eo",
+  "tags": ["product:eo", "domain:connections", "type:data-access"]
+}
