+++ libs/eo/connections/feature-invite/index.ts
@@ -1,0 +1,2 @@
+export * from './invite-connection.component';
+export * from './invite-connection.service';
