+++ libs/eo/connections/feature-overview/index.ts
@@ -1,0 +1,1 @@
+export * from './connections.component';
