+++ libs/eo/connections/shell/index.ts
@@ -1,0 +1,1 @@
+export * from './connections.routes';
